requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy>=1.4.42,<1.5
pydantic==2.6.1
python-dotenv==1.0.0
google-generativeai==0.3.1
pandas==2.2.0
nltk==3.8.1
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0
databases==0.8.0
aiosqlite==0.19.0
